+++ modules/hosts/desktop/default.nix
@@ -7,12 +7,14 @@
 {
   imports = [
     # Include the results of the hardware scan.
+    ./disko.nix
     ./hardware-configuration.nix
   ];
 
   # Bootloader.
   boot.loader.systemd-boot.enable = true;
   boot.loader.efi.canTouchEfiVariables = true;
+  boot.loader.efi.efiSysMountPoint = "/boot";
 
   networking.hostName = "nixos"; # Define your hostname.
   # networking.wireless.enable = true;  # Enables wireless support via wpa_supplicant.